creates test suites for each language family
colocates them with language family samples
