+++ go.mod
@@ -1,9 +1,10 @@
-module github.com/paketo-buildpacks/samples/tests
+module github.com/paketo-buildpacks/samples
 
 go 1.15
 
 require (
 	github.com/docker/distribution v2.7.1+incompatible
+	github.com/gorilla/mux v1.7.3
 	github.com/onsi/gomega v1.10.5
 	github.com/paketo-buildpacks/occam v0.1.0
 	github.com/paketo-buildpacks/packit v0.6.1
