+++ httpd/httpd-sample/httpd.conf
