+++ nginx/nginx_test.go
@@ -1,16 +1,49 @@
-package samples_test
+package nginx_test
 
 import (
+	"flag"
 	"fmt"
 	"os"
 	"testing"
+	"time"
 
 	"github.com/paketo-buildpacks/occam"
+	"github.com/paketo-buildpacks/samples/tests"
 	"github.com/sclevine/spec"
+	"github.com/sclevine/spec/report"
 
 	. "github.com/onsi/gomega"
 	. "github.com/paketo-buildpacks/occam/matchers"
 )
+
+var builders tests.BuilderFlags
+var suite spec.Suite
+
+func init() {
+	flag.Var(&builders, "name", "the name a builder to test with")
+}
+
+func TestNginx(t *testing.T) {
+	Expect := NewWithT(t).Expect
+
+	Expect(len(builders)).NotTo(Equal(0))
+
+	SetDefaultEventuallyTimeout(60 * time.Second)
+
+	suite := spec.New("NGINX", spec.Parallel(), spec.Report(report.Terminal{}))
+	for _, builder := range builders {
+		switch builderType := tests.FindBuilderType(builder); builderType {
+		case "tiny":
+			// do nothing; NGINX does not run on Tiny
+			suite(fmt.Sprintf("NGINX with %s builder", builder), func(t *testing.T, context spec.G, it spec.S) {})
+		case "base":
+			suite(fmt.Sprintf("NGINX with %s builder", builder), testNGINXWithBuilder(builder))
+		default:
+			suite(fmt.Sprintf("NGINX with %s builder", builder), testNGINXWithBuilder(builder))
+		}
+	}
+	suite.Run(t)
+}
 
 func testNGINXWithBuilder(builder string) func(*testing.T, spec.G, spec.S) {
 	return func(t *testing.T, context spec.G, it spec.S) {
@@ -52,7 +85,7 @@ func testNGINXWithBuilder(builder string) func(*testing.T, spec.G, spec.S) {
 			context("app uses nginx", func() {
 				it("builds successfully", func() {
 					var err error
-					source, err = occam.Source("../nginx")
+					source, err = occam.Source("../nginx/nginx-sample")
 					Expect(err).NotTo(HaveOccurred())
 
 					var logs fmt.Stringer
