+++ ruby/ruby_test.go
@@ -1,17 +1,50 @@
-package samples_test
+package ruby_test
 
 import (
+	"flag"
 	"fmt"
 	"os"
 	"path/filepath"
 	"testing"
+	"time"
 
 	"github.com/paketo-buildpacks/occam"
+	"github.com/paketo-buildpacks/samples/tests"
 	"github.com/sclevine/spec"
+	"github.com/sclevine/spec/report"
 
 	. "github.com/onsi/gomega"
 	. "github.com/paketo-buildpacks/occam/matchers"
 )
+
+var builders tests.BuilderFlags
+var suite spec.Suite
+
+func init() {
+	flag.Var(&builders, "name", "the name a builder to test with")
+}
+
+func TestRuby(t *testing.T) {
+	Expect := NewWithT(t).Expect
+
+	Expect(len(builders)).NotTo(Equal(0))
+
+	SetDefaultEventuallyTimeout(60 * time.Second)
+
+	suite := spec.New("Ruby", spec.Parallel(), spec.Report(report.Terminal{}))
+	for _, builder := range builders {
+		switch builderType := tests.FindBuilderType(builder); builderType {
+		case "tiny":
+			// do nothing; Ruby does not run on Tiny
+			suite(fmt.Sprintf("Ruby with %s builder", builder), func(t *testing.T, context spec.G, it spec.S) {})
+		case "base":
+			suite(fmt.Sprintf("Ruby with %s builder", builder), testRubyWithBuilder(builder))
+		default:
+			suite(fmt.Sprintf("Ruby with %s builder", builder), testRubyWithBuilder(builder))
+		}
+	}
+	suite.Run(t)
+}
 
 func testRubyWithBuilder(builder string) func(*testing.T, spec.G, spec.S) {
 	return func(t *testing.T, context spec.G, it spec.S) {
